Update ve33 emission time handling

+++ src/Ve33Periphery.sol
@@ -31,11 +31,11 @@ contract Ve33Periphery is BaseLocker {
     receive() external payable {}
 
     /// @notice Schedules global Ve33 emissions.
-    /// @param startTime Emission schedule start time, or zero for immediate start.
-    /// @param endTime Valid timestamp when the emission stream ends.
+    /// @param startTime Real emission schedule start time, or zero for immediate start.
+    /// @param endTime Valid real timestamp when the emission stream ends.
     /// @param rewardRate Q32 global emission rate in stake tokens per second.
     /// @return amount Amount of stake token required by the schedule.
-    function scheduleEmissions(uint32 startTime, uint32 endTime, uint224 rewardRate) external returns (uint224 amount) {
+    function scheduleEmissions(uint64 startTime, uint64 endTime, uint224 rewardRate) external returns (uint224 amount) {
         amount = abi.decode(
             lock(abi.encode(CALL_TYPE_SCHEDULE_EMISSIONS, msg.sender, startTime, endTime, rewardRate)), (uint224)
         );
@@ -46,8 +46,8 @@ contract Ve33Periphery is BaseLocker {
         uint256 callType = abi.decode(data, (uint256));
 
         if (callType == CALL_TYPE_SCHEDULE_EMISSIONS) {
-            (, address payer, uint32 startTime, uint32 endTime, uint224 rewardRate) =
-                abi.decode(data, (uint256, address, uint32, uint32, uint224));
+            (, address payer, uint64 startTime, uint64 endTime, uint224 rewardRate) =
+                abi.decode(data, (uint256, address, uint64, uint64, uint224));
             uint224 amount = Ve33Lib.scheduleEmissions(CORE_REF, ve33, startTime, endTime, rewardRate);
             result = abi.encode(amount);
             if (amount != 0) ACCOUNTANT.payFrom(payer, stakeToken, amount);
